requirements.txt
openai-whisper>=20231117
moviepy>=1.0.3
numpy>=1.21.0
pillow>=8.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6 